unsigned long -> uint32_t; NFC

+++ clang/include/clang/Lex/PreprocessorOptions.h
@@ -200,7 +200,7 @@ public:
 
   /// The initial value for __COUNTER__; typically is zero but can be set via a
   /// -cc1 flag for testing purposes.
-  unsigned long InitialCounterValue = 0;
+  uint32_t InitialCounterValue = 0;
 
 public:
   PreprocessorOptions() : PrecompiledPreambleBytes(0, false) {}
